feat: f32 HullIntegrity migration + shared apply_hull_damage helper (#155)
Convert HullIntegrity.hp from i32 to f32 end-to-end, making the
wire format (SimSnapshot.hull_integrity, EntityConfig.hull_integrity)
f32 as well. Extract a shared apply_hull_damage helper in damage.rs
that serves as the single code path for applying hull damage and
tracking breakdown bucket crossings. handle_collisions in simulation.rs
now uses this helper instead of duplicating the apply->roll->push
sequence. The client display code (viewscreen_border) rounds the float
to i32 for presentation. breakdowns_from_damage now accepts f32
arguments, preserving all existing integer-threshold behavior while
also supporting fractional accumulation.

Files changed: 11
- src/damage.rs: HullIntegrity f32, new apply_hull_damage helper,
  apply_damage_with_shields no longer applies to hull (returns leakage)
- src/breakdown.rs: breakdowns_from_damage accepts f32
- src/messages.rs: SimSnapshot.hull_integrity: f32
- src/entity_config.rs: HullConfig.hull_integrity: f32
- src/ship_state.rs: snapshot() takes f32 hull_integrity
- src/simulation.rs: handle_collisions uses apply_hull_damage,
  BreakdownQueueResource.cumulative_damage f32, REPAIR_TEAM_HP f32
- src/client_lobby.rs, src/client_sim.rs, src/codec.rs: f32 literals
- src/viewscreen_border.rs: round float for display
- src/config_cache.rs: f32 comparison in test

+++ src/messages.rs
@@ -193,7 +193,7 @@ pub struct SimSnapshot {
     pub ship_x: f32,
     pub ship_z: f32,
     pub ship_yaw: f32,
-    pub hull_integrity: i32,
+    pub hull_integrity: f32,
     /// Current power allocation levels (Helm, Weapons, Science).
     /// Added to SimSnapshot so all clients see the current configuration.
     #[serde(default = "default_power_levels")]
